Make the requirements for loading configs less dumb
This simplifies the loading requirements of the config files. We remove
the dumb layer of pulling custom environment variable keys to interprete
from the config file itself and streamline the process.
`application.ini` is the default file looked for unless defined
otherwise at runtime. `{{keywords}}` are interpolated against environ
values, if matched, on load.  ConfigParser interpolation is disabled as
it causes nothing but issues with tokens and keys which commonly
containt the `%` interpolation character.